+++ src/index.ts
@@ -17,6 +17,7 @@ export {
   rowTestID,
   rowDetailTestID,
   rowSelectTestID,
+  rowFadeTestID,
   selectAllTestID,
 } from './DataTable';